Crypto.Secp256k1.Internal.Base

Description

Crytpographic functions from Bitcoin’s secp256k1 library.

The API for this module may change at any time. This is an internal module only exposed for hacking and experimentation.

Documentation

newtype PubKey

Constructors

PubKey

Fields

get :: ByteString

newtype Msg

Constructors

Msg

Fields

get :: ByteString

newtype Sig

Constructors

Sig

Fields

get :: ByteString

newtype SecKey

Constructors

SecKey

Fields

get :: ByteString

newtype Tweak

Constructors

Tweak

Fields

get :: ByteString

newtype CompactSig

Constructors

CompactSig

Fields

get :: ByteString

sig :: ByteString -> Maybe Sig

Import 64-byte ByteString as Sig.

pubKey :: ByteString -> Maybe PubKey

Import 64-byte ByteString as PubKey.

msg :: ByteString -> Maybe Msg

Import 32-byte ByteString as Msg.

secKey :: ByteString -> Maybe SecKey

Import 32-byte ByteString as SecKey.

compactSig :: ByteString -> Maybe CompactSig

normalizeSig :: Ctx -> Sig -> Maybe Sig

Convert signature to a normalized lower-S form. Nothing indicates that it was already normal.

tweak :: ByteString -> Maybe Tweak

32-Byte ByteString as Tweak.

importPubKey :: Ctx -> ByteString -> Maybe PubKey

Import DER-encoded public key.

exportPubKey :: Ctx -> Bool -> PubKey -> ByteString

Encode public key as DER. First argument True for compressed output.

exportCompactSig :: Ctx -> Sig -> CompactSig

importCompactSig :: Ctx -> CompactSig -> Maybe Sig

importSig :: Ctx -> ByteString -> Maybe Sig

Import DER-encoded signature.

exportSig :: Ctx -> Sig -> ByteString

Encode signature as strict DER.

verifySig :: Ctx -> PubKey -> Sig -> Msg -> Bool

Verify message signature. True means that the signature is correct.

signMsg :: Ctx -> SecKey -> Msg -> Sig

derivePubKey :: Ctx -> SecKey -> PubKey

tweakAddSecKey :: Ctx -> SecKey -> Tweak -> Maybe SecKey

Add tweak to secret key.

tweakMulSecKey :: Ctx -> SecKey -> Tweak -> Maybe SecKey

Multiply secret key by tweak.

tweakAddPubKey :: Ctx -> PubKey -> Tweak -> Maybe PubKey

Add tweak to public key. Tweak is multiplied first by G to obtain a point.

tweakMulPubKey :: Ctx -> PubKey -> Tweak -> Maybe PubKey

Multiply public key by tweak. Tweak is multiplied first by G to obtain a point.

combinePubKeys :: Ctx -> [PubKey] -> Maybe PubKey

Add multiple public keys together.

tweakNegate :: Ctx -> Tweak -> Maybe Tweak