Crypto.Secp256k1.Internal.BaseOps

Description

The API for this module may change at any time. This is an internal module only exposed for hacking and experimentation.

Documentation

ecPubKeyParse

Arguments

:: Ptr LCtx

-> Ptr PubKey64

-> Ptr CUChar
encoded public key array

-> CSize
size of encoded public key array

-> IO Ret

ecPubKeySerialize

Arguments

:: Ptr LCtx

-> Ptr CUChar
array for encoded public key, must be large enough

-> Ptr CSize
size of encoded public key, will be updated

-> Ptr PubKey64

-> SerFlags

-> IO Ret

ecdsaSignatureParseCompact :: Ptr LCtx -> Ptr Sig64 -> Ptr Compact64 -> IO Ret

ecdsaSignatureParseDer

Arguments

:: Ptr LCtx

-> Ptr Sig64

-> Ptr CUChar
encoded DER signature

-> CSize
size of encoded signature

-> IO Ret

ecdsaSignatureSerializeDer

Arguments

:: Ptr LCtx

-> Ptr CUChar
array for encoded signature, must be large enough

-> Ptr CSize
size of encoded signature, will be updated

-> Ptr Sig64

-> IO Ret

ecdsaSignatureSerializeCompact :: Ptr LCtx -> Ptr Compact64 -> Ptr Sig64 -> IO Ret

ecdsaVerify :: Ptr LCtx -> Ptr Sig64 -> Ptr Msg32 -> Ptr PubKey64 -> IO Ret

ecdsaSignatureNormalize

Arguments

:: Ptr LCtx

-> Ptr Sig64
output

-> Ptr Sig64
input

-> IO Ret

ecdsaSign

Arguments

:: Ptr LCtx

-> Ptr Sig64

-> Ptr Msg32

-> Ptr SecKey32

-> FunPtr (NonceFun a)

-> Ptr a
nonce data

-> IO Ret

ecSecKeyVerify :: Ptr LCtx -> Ptr SecKey32 -> IO Ret

ecPubKeyCreate :: Ptr LCtx -> Ptr PubKey64 -> Ptr SecKey32 -> IO Ret

ecSecKeyTweakAdd :: Ptr LCtx -> Ptr SecKey32 -> Ptr Tweak32 -> IO Ret

ecTweakNegate :: Ptr LCtx -> Ptr Tweak32 -> IO Ret

ecPubKeyTweakAdd :: Ptr LCtx -> Ptr PubKey64 -> Ptr Tweak32 -> IO Ret

ecSecKeyTweakMul :: Ptr LCtx -> Ptr SecKey32 -> Ptr Tweak32 -> IO Ret

ecPubKeyTweakMul :: Ptr LCtx -> Ptr PubKey64 -> Ptr Tweak32 -> IO Ret

ecPubKeyCombine

Arguments

:: Ptr LCtx

-> Ptr PubKey64
pointer to public key storage

-> Ptr (Ptr PubKey64)
pointer to array of public keys

-> CInt
number of public keys

-> IO Ret