Crypto.Secp256k1.Internal.Context

Description

The API for this module may change at any time. This is an internal module only exposed for hacking and experimentation.

Documentation

newtype Ctx

Constructors

Ctx

Fields

get :: Ptr LCtx

randomizeContext :: Ctx -> IO ()

createContext :: IO Ctx

cloneContext :: Ctx -> IO Ctx

destroyContext :: Ctx -> IO ()

withContext :: (Ctx -> IO a) -> IO a

verify :: CtxFlags

sign :: CtxFlags

signVerify :: CtxFlags

contextCreate :: CtxFlags -> IO (Ptr LCtx)

contextClone :: Ptr LCtx -> IO (Ptr LCtx)

contextDestroy :: Ptr LCtx -> IO ()

setIllegalCallback

Arguments

:: Ptr LCtx

-> FunPtr (CString -> Ptr a -> IO ())
message, data

-> Ptr a
data

-> IO ()

setErrorCallback

Arguments

:: Ptr LCtx

-> FunPtr (CString -> Ptr a -> IO ())
message, data

-> Ptr a
data

-> IO ()

contextRandomize :: Ptr LCtx -> Ptr Seed32 -> IO Ret