Crypto.Secp256k1.Internal.ForeignTypes

Description

The API for this module may change at any time. This is an internal module only exposed for hacking and experimentation.

Documentation

data LCtx

data PubKey64

data Msg32

data Sig64

data Seed32

data Compact64

data SecKey32

data Tweak32

data RecSig65

type CtxFlags = CUInt

type SerFlags = CUInt

type Ret = CInt

type NonceFun a = Ptr CUChar -> Ptr CUChar -> Ptr CUChar -> Ptr CUChar -> Ptr a -> CInt -> IO CInt

compressed :: SerFlags

uncompressed :: SerFlags

isSuccess :: Ret -> Bool