Crypto.Secp256k1.Internal.Util

Description

The API for this module may change at any time. This is an internal module only exposed for hacking and experimentation.

Documentation

decodeHex :: ConvertibleStrings a ByteString => a -> Maybe ByteString

showsHex :: ByteString -> ShowS

unsafeUseByteString :: ByteString -> ((Ptr a, CSize) -> IO b) -> IO b

useByteString :: ByteString -> ((Ptr a, CSize) -> IO b) -> IO b

unsafePackByteString :: (Ptr a, CSize) -> IO ByteString

packByteString :: (Ptr a, CSize) -> IO ByteString

withRandomSeed :: (Ptr Seed32 -> IO a) -> IO a